스플래쉬 화면 추가 & 해당화면에서 전체마켓 가져오기
